Move interop classes and functions to a separate file
